Update favouriteCommand and related files
Favourite takes a while to be reflected on GUI
and when reopening the GUI, the first favourite
command does not work as intended on a person
who is already added to favourites previously.

This has to change to prevent user dissatisfaction.

Let's make the favourite a BooleanProperty
so that we can actively listen for change and also
create a method in model to retrieve the current
people in favourites from pervious model launch.