Update listCommand
If used on a filtered view based on
find commands, the list command
will not display all the contacts.

Thus, users have no way of going back
to see all their contacts unless if
they relaunch the GUI.

Let's update listCommand to update
model with all the contacts and then
sort it based on the comparator.